// src/types/ChainId.ts
export enum ChainId {
  MAINNET = 1,
  FUEL_MAINNET = 2,
  // LOCAL_TESTNET = 1337,
  // RINKEBY_TESTNET = 4,
  POLYGON_MAINNET = 137,
  OPTIMISM_MAINNET = 10,
  EVMOS_MAINNET = 9001,
}

export enum ChainTag {
  MAINNET = "eth",
  FUEL_MAINNET = "fuel",
  // LOCAL_TESTNET = "localhost",
  // RINKEBY_TESTNET = "rinkeby",
  POLYGON_MAINNET = "polygon",
  OPTIMISM_MAINNET = "optimism",
  EVMOS_MAINNET = "evmos",
}